Disable monthly subfolder toggle while dolphin is open, remove leftover replay settings from dolphin settings tab
